Add LegendRenderer.
